### 2020-01-07 #1999 @whedon
[PRE REVIEW]: thresholdmodeling: A Python package for modeling excesses over a threshold using the Peak-Over-Threshold Method and the Generalized Pareto Distribution
**Submitting author:** @iagolemos1 (<a href="http://orcid.org/0000-0002-5829-7711">Iago Lemos</a>)
**Repository:** <a href="https://github.com/iagolemos1/thresholdmodeling" target ="_blank">https://github.com/iagolemos1/thresholdmodeling</a>
**Version:** v0.0.1
