### 2020-01-07 #1999 @drvinceknight
@iagolemos1 some initial comments after a brief scan through the [pdf](https://github.com/openjournals/joss-papers/blob/joss.01999/joss.01999/10.21105.joss.01999.pdf):

- The `url = ` does not fit on the page
- The paper is a bit on the long side for a joss submission, but this is mainly taken up by graphics. Would these be better suited in documentation with a single minimal example in the paper perhaps (at the moment your documentation in the README is sparse)? The installation instructions do not need to be in the paper either. You can see a list of the criteria for review here: https://openjournals.readthedocs.io/en/latest/review_criteria.html and published papers here: https://joss.theoj.org

Regarding reviewers do you have any suggestions?
