### 2020-01-07 #1999 @iagolemos1
Ok, Dr. Vince. I am going to make the changes. The README leads to the package documentation. Hence, I think it will be better to put the graphics in the README and just some of them in the paper for a little example. I am going to take the installation instructions out too.
For reviewers, I don't have any suggestions.
