### 2020-01-07 #1999 @iagolemos1
Would you have any suggestions of what grpahics I should leave in the paper?
