### 2020-01-07 3dc7be6
Update paper.md
