### 2020-01-07 a2d67f4
Update paper.md
