### 2020-01-07 #1999 @iagolemos1
@drvinceknight
The [README](https://github.com/iagolemos1/thresholdmodeling/blob/master/README.md) was updated.
