### 2020-01-08 #1999 @iagolemos1
Answers in  https://github.com/openjournals/joss-reviews/issues/1999.

Em ter., 7 de jan. de 2020 às 13:39, whedon <notifications@github.com>
escreveu:

> 👉 Check article proof 📄 👈
> <https://github.com/openjournals/joss-papers/blob/joss.01999/joss.01999/10.21105.joss.01999.pdf>
>
