### 2020-01-08 #1999 @drvinceknight
@kellieotto @galenstocking @bahung would you be able to review this paper on a piece of software for extreme value analysis? There is information about what that involves here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and I'm happy to answer any questions you may have.
