### 2020-01-08 395af0a
Update Functions Documentation.md
