### 2020-01-09 #1999 @iagolemos1
I did some minimal changes in the citations on the paper.
