### 2020-01-15 9048ac7
Update paper.md
