### 2020-01-16 iagolemos1/thresholdmodeling#1 @kellieotto
Paper grammar
This is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/2013).

I recognize that English might not be your native language. Unfortunately, there are some sentences that are grammatically incorrect and I don't understand the meaning. Below, I will make corrections where I can and point out places where I don't understand the intended meaning. I hope this is helpful!

**first paragraph:** "dealig" --> "dealing", "asymtopic" --> "asymptotic"
**first paragraph:** the last sentence needs to be rewritten as it is not grammatically correct

**second paragraph:** "since the threshold selection until deep statistical analyses" -- I don't know what this means; I think "until" and "since" are the wrong words.

**third paragraph:** "which ones provide" --> "that are used for" ; "which one does not contemple" --> "which does not contain" ; "excesses" --> "many" (?)

**Model Checking section:** "heoretical" --> "theoretical" ; "provided,using" --> "provided, using"

**Declustering section:** "giving empirical rule" --> "given empirical rule" ; "maximuns" --> "maximums"

**Reproducibility and User's Guide:** "confronting" --> "testing"
