### 2020-01-17 iagolemos1/thresholdmodeling#1 @iagolemos1
Thanks, @kellieotto. I am going to consider all the corrections. Some of the errors are lack of attention. I am going to try to rewrite the sentences with grammatical errors.
