### 2020-01-17 2e75678
Update paper.md
