### 2020-01-19 iagolemos1/thresholdmodeling#2 @bahung
Paper grammar
This is part of your [paper reviewing process](https://github.com/openjournals/joss-reviews/issues/2013#issuecomment-575421666) related to grammar checking.

While the paper has rich documentation and functionality, its document needs to be written in a more concise manner to make its claim/objectives clear for readers/testers. I list some confusing points below which refrain me from doing further reviewing step.

### In summary

1. The last sentence in the 1st paragraph is not clear and has extra punctuation: "... in an unusually level, being of minimum or maximum.."

2. 3rd paragraph: Sentence "Otherwise, in Python, it is possible to find the
scikit-extremes (Correoso, 2019), which one does not contain the threshold models yet." I think "one" does not fit here.

3. Also in 3rd paragraph: "this approach" in "Moreover, this package brings to the community
of scientists, engineers and any other interested person and programmer the possibility to
conduct an extreme value analysis using a strong, consolidated and high-level programming
language given the importance of this approach in corrosion engineering" should be more specific.

### In Package Features

4. **Parameter Stability Plot**: "Also, it is possible to get the two graphics related to the shape and the modified scale parameters stability plot" this sentence is not clear in "modified scale parameters stability plot".

5. **Probability Density Function, Comulative Distribution Function, QuantileQuantile and Probability-Probability Plots:**  grammar error: comulative --> cumulative

6. The paragraph of **Probability Density Function, Comulative Distribution Function, QuantileQuantile and Probability-Probability Plots:** needs rephrasing to make its sentences clear.
