### 2020-01-19 iagolemos1/thresholdmodeling#2 @iagolemos1
@bahung Thanks very much for this. I am going to consider everything in order to make the paper even better.
