### 2020-01-19 ccd3670
Update paper.md
