### 2020-01-19 8f78b06
Update paper.md
