### 2020-01-19 301377c
Update paper.md
