### 2020-01-19 a3af2bc
Update paper.md
