### 2020-01-19 iagolemos1/thresholdmodeling#2 @iagolemos1
@bahung I tried to consider everything you said. I put an example in the first paragraph and tried to make it clear, the same in the 3rd paragraph with the 'approach' problem. Feel free to sugger how to make any change, please, it would help me a lot.
