### 2020-01-19 #2013 @iagolemos1
Below, I tried to consider all the changes suggested by @bahung.
