### 2020-01-19 f2c2dd8
Update paper.md
