+++ README.md
@@ -9,7 +9,7 @@ In this repository you can find the main files of the package, the [Functions Do
 
 * For installing Anaconda on Linux, go to [this link](https://docs.anaconda.com/anaconda/install/linux/). For installing on Windows, go to [this one](https://docs.anaconda.com/anaconda/install/windows/). 
 
-* For create your own environment, see [here](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-with-commands). For Linux it uses the terminal and Windows it uses the Anaconda Prompt. 
+* For creating your own environment, see [here](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#creating-an-environment-with-commands). For Linux it uses the terminal and Windows it uses the Anaconda Prompt. 
 
 ## Windows Users 
 By using Anaconda Navigator, install R on your environment (easier for Windows users).  
