### 2020-01-24 iagolemos1/thresholdmodeling#4 @kellieotto
Cannot reproduce first example
This is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/2013).

I was able to reproduce all the code examples in the README except for the first one. I get this error:

```
In [17]: thresh_modeling.Parameter_Stability_plot(data, 0.05)
    ...:
/opt/conda/lib/python3.7/site-packages/rpy2/rinterface_lib/conversion.py in _get_cdata(obj)
    166         try:
--> 167             cdata = obj.__sexp__._cdata
    168         except AttributeError:

AttributeError: 'numpy.float64' object has no attribute '__sexp__'

<ipython-input-17-4cb15b247b71> in <module>
----> 1 thresh_modeling.Parameter_Stability_plot(data, 0.05)

/opt/conda/lib/python3.7/site-packages/thresholdmodeling/thresh_modeling.py in Parameter_Stability_plot(sample, alpha)
    101     #Getting parameters and CI's for both plots
    102     for u in threshold:
--> 103         fit = POT.fitgpd(rdata, u, est = 'mle')  #fitting distribution using POT package with the MLE method
    104         shape.append(fit[0][1]) #adding the shape parameter to the respective array
    105         scale.append(fit[0][0]) #adding the scale parameter to the respective array

/opt/conda/lib/python3.7/site-packages/rpy2/robjects/functions.py in __call__(self, *args, **kwargs)
    190                 kwargs[r_k] = v
    191         return (super(SignatureTranslatedFunction, self)
--> 192                 .__call__(*args, **kwargs))
    193
    194

/opt/conda/lib/python3.7/site-packages/rpy2/robjects/functions.py in __call__(self, *args, **kwargs)
    119             else:
    120                 new_kwargs[k] = conversion.py2rpy(v)
--> 121         res = super(Function, self).__call__(*new_args, **new_kwargs)
    122         res = conversion.rpy2py(res)
    123         return res

/opt/conda/lib/python3.7/site-packages/rpy2/rinterface_lib/conversion.py in _(*args, **kwargs)
     38 def _cdata_res_to_rinterface(function):
     39     def _(*args, **kwargs):
---> 40         cdata = function(*args, **kwargs)
     41         # TODO: test cdata is of the expected CType
     42         return _cdata_to_rinterface(cdata)

/opt/conda/lib/python3.7/site-packages/rpy2/rinterface.py in __call__(self, *args, **kwargs)
    780             call_r = rmemory.protect(
    781                 _rinterface.build_rcall(self.__sexp__._cdata, args,
--> 782                                         kwargs.items()))
    783             res = rmemory.protect(
    784                 openrlib.rlib.R_tryEval(

/opt/conda/lib/python3.7/site-packages/rpy2/rinterface_lib/_rinterface_capi.py in build_rcall(rfunction, args, kwargs)
    225         item = rlib.CDR(rcall)
    226         for val in args:
--> 227             cdata = rmemory.protect(conversion._get_cdata(val))
    228             rlib.SETCAR(item, cdata)
    229             item = rlib.CDR(item)

/opt/conda/lib/python3.7/site-packages/rpy2/rinterface_lib/conversion.py in _get_cdata(obj)
    168         except AttributeError:
    169             raise ValueError('Not an rpy2 R object and unable '
--> 170                              'to cast it into one: %s' % repr(obj))
    171     else:
    172         cdata = cast(obj)

ValueError: Not an rpy2 R object and unable to cast it into one: 0.0
```
