### 2020-01-24 iagolemos1/thresholdmodeling#4 @iagolemos1
Did you try exactly what I presented? Another error that I don't know what's about. I am going to try to find out.
