### 2020-01-24 iagolemos1/thresholdmodeling#4 @iagolemos1
What's your rpy2 version?
Run this on your IDE:
```
import rpy2
print(rpy2.__version__)
```
