### 2020-01-24 iagolemos1/thresholdmodeling#4 @kellieotto
```
[18]: import rpy2
    ...: print(rpy2.__version__)
3.2.4
```
