### 2020-01-24 iagolemos1/thresholdmodeling#4 @kellieotto
Yes, I copied and pasted the code exactly as it is written in the README.
