### 2020-01-24 iagolemos1/thresholdmodeling#4 @iagolemos1
I think it's the version of rpy2. Windows users runs that with rpy2 version
2.9.4.
Uninstall the existing version and use the one from Anaconda repository.

Em sex, 24 de jan de 2020 16:38, Kellie Ottoboni <notifications@github.com>
escreveu:

> [18]: import rpy2
>     ...: print(rpy2.__version__)
> 3.2.4
>
