### 2020-01-24 iagolemos1/thresholdmodeling#4 @iagolemos1
The problem is the same from the previous topic: I wrote the code in Windows using the 2.9.4 version (the last version for Windows, from the Anaconda Repository). The setup.py of the package installs rpy2 from PyPi repository (it's ok for Linux users, but for Windows ones, a error will be given, like @bahung problem). Unfortunately, I did not know about this. I will fix everything on README for Linux Users and in a next release I will take rpy2 out from the setup.py and everyone will install it mannualy.
