### 2020-01-24 iagolemos1/thresholdmodeling#4 @iagolemos1
just run:
```
pip uninstall rpy2
conda install -c r rpy2
```
