### 2020-01-24 iagolemos1/thresholdmodeling#4 @kellieotto
> just run:
>
> ```
> pip uninstall rpy2
> conda install -c r rpy2
> ```

Actually, I ran this. Now the example works.
```
conda install -c r rpy2=2.9.4
```

Please add this to the README installation instructions so that people can install a version that works with the software.
