### 2020-01-24 iagolemos1/thresholdmodeling#1 @kellieotto
The paper is much improved. A few more small grammar edits:

**first paragraph**: "been emerged"-->"emerged" ; "maximums observations"-->"maximum observations"

**second paragraph**: "extremes of dependences sequences" I'm not sure exactly what this phrase means.

**third paragraph**: "into a software context" --> "In software"
