### 2020-01-24 iagolemos1/thresholdmodeling#5 @kellieotto
Testing
This is part of your [JOSS review](https://github.com/openjournals/joss-reviews/issues/2013).

[This file is called test.py](https://github.com/iagolemos1/thresholdmodeling/blob/master/Test/test.py), but these aren't unit tests. Unit tests check that the code is producing the output that you expect. [This site](https://realpython.com/python-testing/) has detailed examples on how to write tests.

The expectation from JOSS is this: Are there automated tests or manual steps described so that the functionality of the software can be verified?
