### 2020-01-25 iagolemos1/thresholdmodeling#5 @iagolemos1
@kellieotto The file test is just something I used to test the functions. The main idea for the expectation from JOSS is in the README, where I describe, with steps, the funcionality of the software. Isn't that enough? because if it not, I will provide an automated test.
