### 2020-01-25 iagolemos1/thresholdmodeling#1 @iagolemos1
I made the changes.
