### 2020-01-25 iagolemos1/thresholdmodeling#5 @iagolemos1
@kellieotto I am just confused with unit testing. The functions from ``thresholdmodeling`` are quite unpredictable, I mean, for example, how am I going to write a unittest for the Parameter_Stability_Function if it uses the log-likelihood function several times with reparameterization of an estimated parameter in order to draw a graphic? The only way I see it is taking the answer a calling the test for the given data and compare it to the answer. Please, help me to clarify this :P
