### 2020-01-25 iagolemos1/thresholdmodeling#5 @iagolemos1
Now I understand what it supposed to be. I don't know exactly what I should do. The package functions are quite unpredictable, I mean, it is hard to test what functions that will compute estimatives of the log-likelihood function several times to give an answer will return, like the Parameter_Stability_Function.
When I wrote the README, the idea was to give the answers presented in the Coles's book. I mean, they are manual steps, right? If anyone goes to the book, the answers in the README can be verified, and if anyone reproduce the steps and get that results, the software functionality can be also verified.
