### 2020-01-25 iagolemos1/thresholdmodeling#5 @iagolemos1
Also, there are some functions that we can't verify by Coles's book. But we can verify that using another packages from R, for example. I already did that when I was writing. What do you think about that?
