### 2020-01-25 079a4a2
Delete test.py
