### 2020-01-25 661e04f
Update examples.py
