+++ Functions Examples/examples.py
@@ -21,4 +21,4 @@ thresh_modeling.lmom_dist(data, 30, 'mle')
 thresh_modeling.lmom_sample(data)
 thresh_modeling.lmomplot(data, 30)
 thresh_modeling.decluster(data, 30, 30)
-thresh_modeling.entropy(data, 2, 30, 'mle')
+thresh_modeling.entropy(data, 'e', 30, 'mle')
