+++ .travis.yml
@@ -1,0 +1,7 @@
+python:
+  - "2.7"
+  - "3.7"
+install:
+  - pip install -r requirements.txt
+script:
+  - python -m unittest discover
