+++ .travis.yml
@@ -1,5 +1,4 @@
 python:
-  - "2.7"
   - "3.7"
 install:
   - pip install -r requirements.txt
