+++ .travis.yml
@@ -1,4 +1,6 @@
+language: python
 python:
+  - "2.7"
   - "3.7"
 install:
   - pip install -r requirements.txt
