+++ requirements.txt
@@ -2,6 +2,5 @@ scipy
 numpy 
 pandas 
 thresholdmodeling 
-math
 matplotlib
 rpy2==2.9.4 
