+++ requirements.txt
@@ -3,4 +3,4 @@ numpy
 pandas 
 thresholdmodeling 
 matplotlib
-rpy2==2.9.4 
+
