### 2020-01-25 54d6c16
Update .travis.yml
