+++ .travis.yml
@@ -4,6 +4,7 @@ python:
   - "3.7"
 install:
   - pip install -r requirements.txt
+  - conda install r
   - conda install -c r rpy2==2.9.4
 script:
   - python -m unittest discover
