+++ .travis.yml
@@ -5,6 +5,6 @@ python:
 install:
   - pip install -r requirements.txt
   - conda install r
-  - conda install -c r rpy2==2.9.4
+  - conda install -c r rpy2
 script:
   - python -m unittest discover
