### 2020-01-25 c6357d4
Update .travis.yml
