+++ .travis.yml
@@ -3,7 +3,7 @@ python:
   - "2.7"
   - "3.7"
 install:
-  - pip install -r requirements.txt
+  - pip install thresholdmodeling
   - conda install r
   - conda install -c r rpy2=2.9.4
 script:
