### 2020-01-25 185a3e5
Update .travis.yml
