+++ .travis.yml
@@ -3,8 +3,9 @@ python:
   - "2.7"
   - "3.7"
 install:
-  - pip install thresholdmodeling
   - conda install r
   - conda install -c r rpy2=2.9.4
+  - pip install thresholdmodeling
+  
 script:
   - python -m unittest discover
