### 2020-01-25 496316f
Update .travis.yml
