+++ .travis.yml
@@ -1,11 +1,28 @@
 language: python
 python:
+  # We don't actually use the Travis Python, but this keeps it organized.
   - "2.7"
-  - "3.7"
+  - "3.5"
+  - "3.6"
 install:
-  - conda install r
-  - conda install -c r rpy2=2.9.4
-  - pip install thresholdmodeling
+  - sudo apt-get update
+  # We do this conditionally because it saves us some downloading if the
+  # version is the same.
+  - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
+      wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
+    else
+      wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+    fi
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - source "$HOME/miniconda/etc/profile.d/conda.sh"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
+  # Useful for debugging any issues with conda
+  - conda info -a
+
+  # Replace dep1 dep2 ... with your dependencies
+  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION rpy2 r ...
+  - conda activate test-environment
+  - python setup.py install
   
-script:
-  - python -m unittest discover
