+++ .travis.yml
@@ -22,7 +22,7 @@ install:
   - conda info -a
 
   # Replace dep1 dep2 ... with your dependencies
-  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION rpy2 r ...
+  - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION rpy2 r 
   - conda activate test-environment
   - python setup.py install
   
