### 2020-01-25 b0cb67d
Update .travis.yml
