### 2020-01-25 42c1123
Update .travis.yml
