+++ .travis.yml
@@ -22,5 +22,7 @@ install:
   # Replace dep1 dep2 ... with your dependencies
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION 
   - conda activate test-environment
+  - conda install r
+  - contal install -c r rpy2=2.9.4
   - python setup.py install
   
