+++ .travis.yml
@@ -23,6 +23,6 @@ install:
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION 
   - conda activate test-environment
   - conda install r
-  - contal install -c r rpy2=2.9.4
+  - conda install -c r rpy2=2.9.4
   - python setup.py install
   
