+++ .travis.yml
@@ -26,3 +26,6 @@ install:
   - conda install -c r rpy2=2.9.4
   - python setup.py install
   
+ script:
+  - python -m unittest discover
+  
