### 2020-01-25 aaac023
Delete non_central_moments_test.py
