### 2020-01-25 c7f772d
Delete lmom_sample_test.py
