### 2020-01-25 320b337
Delete lmom_dist_test.py
