### 2020-01-25 3159c5e
Delete entropy_test.py
