### 2020-01-25 718501e
Update .travis.yml
