+++ .travis.yml
@@ -27,5 +27,5 @@ install:
   - python setup.py install
   
  script:
-  - python -m unittest discover
+  -python setup.py test
   
