### 2020-01-25 ca051c7
Update .travis.yml
