+++ tests/unit/__init__.py
@@ -1,8 +1,1 @@
 
-from rpy2.robjects.packages import importr
-import rpy2.robjects.packages as rpackages
-
-base = importr('base')
-utils = importr('utils')
-utils.chooseCRANmirror(ind=1)
-utils.install_packages('POT') #installing POT package
