### 2020-01-25 1482020
Update declustering_test.py
