### 2020-01-25 cef63a0
Update thresh_modeling.py
