### 2020-01-25 iagolemos1/thresholdmodeling#5 @iagolemos1
@kellieotto while I wait for you answer for the plots, I wrote some unittest codes for numerical values. I ran this in Travis CI but I having problems because it's not running at test. I am receving this at the final:
![image](https://user-images.githubusercontent.com/57317828/73126715-be619f80-3f94-11ea-947a-7be5e3ad1fee.png)
I needed to made a little change in the source code to make it works.
