### 2020-01-25 5bd66a5
Delete return_value_test.py
