### 2020-01-25 6bacb1c
Update setup.py
