+++ setup.py
@@ -13,7 +13,7 @@ setuptools.setup(
     long_description_content_type="text/markdown",
     url="https://github.com/iagolemos1/thresholdmodeling",
     packages=['thresholdmodeling', 'tests'],
-    test_suit = 'tests
+    test_suit = 'tests'
     classifiers=[
         "Programming Language :: Python :: 3",
         'License :: OSI Approved :: GNU General Public License (GPL)',
