+++ .travis.yml
@@ -27,4 +27,5 @@ install:
   - python setup.py install
   
 script:
-  - python tests
+  - python setup.py test
+
