### 2020-01-25 iagolemos1/thresholdmodeling#5 @iagolemos1
Now the automated tests are running as it should be.
![image](https://user-images.githubusercontent.com/57317828/73128784-58384500-3fb3-11ea-9629-f221666bf04d.png)
