### 2020-01-26 iagolemos1/thresholdmodeling#5 @kellieotto
@iagolemos1 The unit tests you wrote in the tests/ directory look good! I understand you can't write tests for plots, but checking the numeric values are what you expect from the Cole examples is good 👍
