### 2020-01-28 iagolemos1/thresholdmodeling#2 @iagolemos1
@bahung Could you give me a feedback about all the changes? :D
