### 2020-02-04 #2013 @bahung
Hi @drvinceknight,

I thought you have approved as you suggested with @kellieotto in [the previous comment](https://github.com/openjournals/joss-reviews/issues/2013#issuecomment-578671696)

TBH, I do not fully understand how do you organise this review by saying that way, so I stop taking further review steps, despite it seems to fail to initiate the package without R_USER and R_HOME being defined in the system paths.
