### 2020-02-04 #2013 @iagolemos1
Hi @bahung, I think it happens when rpy2 package was not successfully or maybe you are having problems with which environment you are using. Just let me know what is happening then we can solve the problem :D
The package should work successfully like it was for @kellieotto.
