### 2020-02-04 #2013 @iagolemos1
I recommend uninstalling everything you did, then follow the steps In the [README](https://github.com/iagolemos1/thresholdmodeling).
```
conda activate my_env
conda install r
conda install -c r rpy2=2.9.4
pip install thresholdmodeling
```
Now I am realizing, this problem you had must be related to the R base for the software.
