### 2020-02-05 #2013 @bahung
> > I thought you have approved as you suggested with @kellieotto in [the previous comment](https://github.com/openjournals/joss-reviews/issues/2013#issuecomment-578671696)
>
> Sorry I think there was a typo in that comment, I meant to write: `we are now waiting on @bahung to finish their review?`.
>
> > TBH, I do not fully understand how do you organise this review by saying that way, so I stop taking further review steps, despite it seems to fail to initiate the package without R_USER and R_HOME being defined in the system paths.
>
> Apologies I do not understand. If you're having problems installing the software this would be useful to know in your review.
>
> We do need two reviews, if you're able to complete the review that would be great but if not please let me know and I can search for another reviewer.

No problem, sorry for the misunderstanding. I will proceed as usual.
