### 2020-02-10 #2013 @drvinceknight
@iagolemos1 some minor typos: https://github.com/iagolemos1/thresholdmodeling/pull/7
