### 2020-02-10 #2013 @iagolemos1
@drvinceknight Now I have to archive the repository in Zenodo, right?
