### 2020-02-10 #2013 @drvinceknight
> @drvinceknight Now I have to archive the repository in Zenodo, right?

That's right: and pleas report the DOI in the review thread. Could you make sure the Zenodo deposit has the correct metadata (title and author list).

Can you also confirm the version of the software please?
