### 2020-02-10 #2013 @danielskatz
👋 @iagolemos1  - please merge https://github.com/iagolemos1/thresholdmodeling/pull/8 or let me know what parts of it you disagree with.  Also, there is at least one case error in the .bib file, a pareto that should be {P}areto.  Please check the bib entries as built in the pdf carefully, and make any other needed changes as well.
