+++ paper.bib
@@ -1,6 +1,6 @@
 @Book{coles,
   author = {S. Coles},
-  title = {An Introduction to Statistical Modeling of Extreme Values},
+  title = {An {I}ntroduction to {S}tatistical {M}odeling of {E}xtreme {V}alues},
   year = {2001},
   edition = {1st},
   publisher = {Springer},
@@ -8,21 +8,21 @@
   doi = {10.1007/978-1-4471-3675-0},
 }
 @Manual{POT,
-	title = {\pkg{POT}: Generalized Pareto Distribution and Peaks Over Threshold},
+	title = {\pkg{POT}: Generalized {P}areto {D}istribution and {P}eaks {O}ver {T}hreshold},
 	author = {Mathieu Ribatet and Christophe Dutang},
 	year = {2019},
 	note = {\proglang{R} package version 1.1-7},
 	url = {https://cran.r-project.org/web/packages/POT/index.html},
 }
 @Manual{extremes,
-	title = {\pkg{extRemes}: Extreme Value Analysis},
+	title = {\pkg{extRemes}: {E}xtreme {V}alue {A}nalysis},
 	author = {Eric Gilleland},
 	year = {2019},
 	note = {\proglang{R} package version 2.0-11},
 	url = {https://cran.r-project.org/web/packages/extRemes/index.html},
 }
 @Manual{evd,
-	title = {\pkg{evd}: AFunctions for Extreme Value Distributions},
+	title = {\pkg{evd}: Functions for {E}xtreme {V}alue {D}istributions},
 	author = {Alec Stephenson},
 	year = {2018},
 	note = {\proglang{R} package version 2.3-3},
@@ -30,7 +30,7 @@
 }
 
 @Manual{ismev,
-  title = {\pkg{ismev}: An Introduction to Statistical Modeling of Extreme Values},
+  title = {\pkg{ismev}: An {I}ntroduction to {S}tatistical {M}odeling of {E}xtreme {V}alues},
   author = {Janet E. Heffernan and Alec G. Stephenson},
   year = {2018},
   note = {\proglang{R} package version 1.42},
@@ -39,8 +39,8 @@
 
 @thesis{tan,
 	author = {Hwei-Yang Tan},
-	title = {Analysis of Corrosion Data for
-		Integrity Assessments},
+	title = {Analysis of {C}orrosion {D}ata for
+		{I}ntegrity {A}ssessments},
 	type = {Thesis for the Degree of Doctor of Philosophy},
 	year = {2017},
 	institution = {Brunel University London},
@@ -49,15 +49,15 @@
 
 @unpublished{esther,
 	author = {Esther Bommier},
-	title = {Peaks-Over-Threshold Modelling of
-	Environmental Data},
+	title = {Peaks-{O}ver-{T}hreshold {M}odelling of
+	{E}nvironmental {D}ata},
 	note = {Examensarbete i matematik, Uppsala University},
 	year = {2014},}
 
 @unpublished{max,
 	author = {Max Rydman},
-	title = {Application of the Peaks-Over-Threshold
-	Method on Insurance Data},
+	title = {Application of the {P}eaks-{O}ver-{T}hreshold
+	{M}ethod on {I}nsurance {D}ata},
 	note = {Examensarbete i matematik, Uppsala University},
 	year = {2018},}
 
@@ -74,7 +74,7 @@
 
 @Book{hosking,
   author = {J. R. M. Hosking and J. R. Wallis},
-  title = {Regional Frequency Analysis: An Approach Based on L-Moments.},
+  title = {Regional {F}requency {A}nalysis: {A}n {A}pproach {B}ased on {L}-{M}oments.},
   year = {1997},
   edition = {1st},
[…]
