### 2020-02-10 #2013 @iagolemos1
Hi @danielskatz. I merged the changes and made the changes in the .bib file. Thanks!
