### 2020-02-10 #2013 @iagolemos1
@danielskatz I tryied to reach the paper using the DOI and it's appearing "File not foud". It's because the DOI isn't ready yet?
![image](https://user-images.githubusercontent.com/57317828/74181061-527f6800-4c1f-11ea-99f8-0f5a7b46f310.png)
