### 2020-02-10 #2013 @danielskatz
The link https://doi.org/10.21105/joss.02013 works fine for me (and did before I closed the issue) - it might just be an issue of how new DOIs resolve and how this information is distributed.  Please try again in a little bit and if it continue to not work for you, let us know.

As I look at this more carefully, the DOI is resolving for you, and bringing you to the right page.  The problem you are having is that the PDF at that page is not displaying, which is likely a web caching issue.  Again, however, waiting a little bit should work.
